2 commit. Postman test passed

+++ src/main/java/ru/yandex/practicum/filmorate/service/FilmService.java
@@ -9,20 +9,21 @@ import ru.yandex.practicum.filmorate.controller.FilmController;
 import ru.yandex.practicum.filmorate.exception.ValidationException;
 import ru.yandex.practicum.filmorate.model.Film;
 import ru.yandex.practicum.filmorate.storage.film.FilmStorage;
+import ru.yandex.practicum.filmorate.storage.user.UserStorage;
 
 import java.time.LocalDate;
-import java.util.Comparator;
 import java.util.List;
-import java.util.stream.Collectors;
 
 @Service
 public class FilmService {
 
     private FilmStorage filmStorage;
+    private UserStorage userStorage;
 
     @Autowired
-    public FilmService(FilmStorage filmStorage) {
+    public FilmService(FilmStorage filmStorage, UserStorage userStorage) {
         this.filmStorage = filmStorage;
+        this.userStorage = userStorage;
     }
     private static final Logger log = LoggerFactory.getLogger(FilmController.class);
 
@@ -36,31 +37,52 @@ public class FilmService {
         filmStorage.addFilm(film);
         return film;
     }
+    
+    public Film getFilm(Integer filmId) {
+        idValidator(filmId);
+        if(filmStorage.containsFilm(filmId)){
+            return filmStorage.getFilm(filmId);
+        }
+        else{
+            throw new NullPointerException("такого id нет");
+        }
+    }
 
     public Film updateFilm(Film film){
         validator(film);
         if(filmStorage.containsFilm(film.getId())){
             log.info("фильм {} обновлен", film.getName());
             filmStorage.updateFilm(film);
-        }else{
+        }
+        else{
             log.error("NOT SUCH FILM TO UPDATE");
-            throw new ValidationException("такого id нет");
+            throw new NullPointerException("такого id нет");
         }
         return film;
     }
 
-    public List<Film> topFilms(int count){
+    public List<Film> topFilms(Integer count){
         return filmStorage.topFilms(count);
     }
 
 
-    public void putLike(int filmId, int userId){
-        filmStorage.putLike(filmId, userId);
+    public void putLike(Integer filmId, Integer userId){
+        idValidator(filmId);
+        if(userStorage.containsUser(userId)) {
+            filmStorage.putLike(filmId, userId);
+        } else {
+            throw new NullPointerException("Пользователя с таким id нет");
+        }
     }
 
 
-    public void deleteLike(int filmId, int userId) {
-        filmStorage.deleteLike(filmId, userId);
+    public void deleteLike(Integer filmId, Integer userId) {
+        idValidator(filmId);
+        if(userStorage.containsUser(userId)) {
+            filmStorage.deleteLike(filmId, userId);
+        } else {
+            throw new NullPointerException("Пользователя с таким id нет");
+        }
     }
 
     private void validator(Film film){
@@ -79,6 +101,13 @@ public class FilmService {
         if(film.getDuration() < 0) {
             log.error("отрицательное число");
             throw new ValidationException("Длительность должна быть положительной");
+        }
+    }
+
+    private void idValidator(Integer id) {
+        if(id < 1) {
+            log.error("неверный id");
+            throw new NullPointerException("id не может быть пустым или отрицательным");
         }
     }
